==> notebooks/01_load_and _validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """
    Download a specific month's raw ride data from a given URL and save it as a .parquet file.
    Parameters:
    - year (int): The year of the desired data.
    - month (int): The month of the desired data.
    Returns:
    - Path: The path to the downloaded .parquet file.
    Raises:
    - Exception: If the URL does not return a 200 status code.
    """
    # Construct the URL using the given year and month
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"

    # Send a GET request to the URL
    response = requests.get(URL)

    # If the response is successful (status code 200)
    if response.status_code == 200:
        # Define the path where the .parquet file will be saved
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        
        # Write the content of the response to the .parquet file
        open(path, "wb").write(response.content)
        
        return path
    else:
        # Raise an exception if the URL is not available
        raise Exception(f"{URL} is not available")


download_one_file_of_raw_data(year=2022, month=1)

import pandas as pd 

rides = pd.read_parquet("../data/raw/rides_2022-01.parquet")

rides.head(20)

# Filter the DataFrame to retain only the 'tpep_pickup_datetime' and 'PULocationID' columns
rides = rides[["tpep_pickup_datetime", "PULocationID"]]

# Rename the relevant columns for better readability
rides.rename(columns={
    "tpep_pickup_datetime": "pickup_datetime",  # Rename 'tpep_pickup_datetime' to 'pickup_datetime'
    "PULocationID": "pickup_location_id",       # Rename 'PULocationID' to 'pickup_location_id'
},inplace=True)

rides.head(20)

rides["pickup_datetime"].describe()

# Filter the rides DataFrame to only include rows where 'pickup_datetime' is on or after "2022-01-01"
rides = rides[rides.pickup_datetime >= "2022-01-01"]

# Further filter the rides DataFrame to only include rows where 'pickup_datetime' is before "2022-02-01"
rides = rides[rides.pickup_datetime < "2022-02-01"]

rides["pickup_datetime"].describe()

# Save the filtered and transformed rides DataFrame to a .parquet file for further use or analysis
rides.to_parquet("../data/transformed/validated_rides_2022-01.parquet")
