==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd

rides = pd.read_parquet("../data/transformed/validated_rides_2022-01.parquet")

rides.head(10)

# Extract the hour from the 'pickup_datetime' column and floor it to get the beginning of the hour.
# This will help in grouping or analyzing the data based on hourly intervals.
rides["pickup_hour"] = rides["pickup_datetime"].dt.floor("H")
rides

# Group the rides DataFrame by the 'pickup_hour' and 'pickup_location_id' columns.
# Count the number of occurrences (rides) for each combination of 'pickup_hour' and 'pickup_location_id'.
# This helps in understanding ride patterns for each location at hourly intervals.
agg_rides = rides.groupby(["pickup_hour", "pickup_location_id"]).size().reset_index()

# Rename the column containing ride counts to "rides"
agg_rides.rename(columns={0: "rides"}, inplace=True)
agg_rides

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    """
    Fill in missing hourly slots for each pickup location with zero rides.
    Parameters:
    - agg_rides (pd.DataFrame): DataFrame with aggregated ride counts.
    Returns:
    - pd.DataFrame: DataFrame with missing hourly slots filled.
    """
    
    # Get unique location IDs
    location_ids = agg_rides["pickup_location_id"].unique()

    # Create a full hourly range based on the min and max pickup hours
    full_range = pd.date_range(
        agg_rides["pickup_hour"].min(), 
        agg_rides["pickup_hour"].max(), freq="H"
    )
    
    output = pd.DataFrame()

    # Loop through each location ID to fill missing hours
    for location_id in tqdm(location_ids):
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]

        agg_rides_i.set_index("pickup_hour", inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)

        agg_rides_i["pickup_location_id"] = location_id

        output = pd.concat([output, agg_rides_i])

    # Reset index and rename columns for the final output
    output = output.reset_index().rename(columns={"index": "pickup_hour"})

    return output


agg_rides_all_slots = add_missing_slots(agg_rides=agg_rides)

from typing import Optional, List
import plotly.express as px

def plot_rides(rides: pd.DataFrame, locations: Optional[List[int]] = None):
    """
    Plot ride counts over time for specified pickup locations.
    Parameters:
    - rides (pd.DataFrame): DataFrame containing ride counts.
    - locations (Optional[List[int]]): List of pickup location IDs to plot. If None, all locations are plotted.
    """
    
    # Filter rides based on provided location IDs or use all if none provided
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides
    
    # Create a line plot using plotly
    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color= "pickup_location_id",
        template="none",
    )

    fig.show()

plot_rides(agg_rides_all_slots, locations=[43])

# Save the aggregated and transformed ride data to a parquet file for further use.
agg_rides_all_slots.to_parquet("../data/transformed/ts_data_2022_01.parquet")
