==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

from datetime import datetime, timedelta
import pandas as pd

current_date = pd.to_datetime(datetime.utcnow()).floor('H')
print(f'{current_date=}')

from src.inference import load_batch_of_features_from_store

features = load_batch_of_features_from_store(current_date=current_date)

from src.inference import load_model_from_registry, get_model_predictions

model = load_model_from_registry()
predictions = get_model_predictions(model, features)

predictions['pickup_hour'] = current_date
predictions

from src.feature_store_api import get_feature_store, FEATURE_GROUP_PREDICTIONS_METADATA

# Connect to the feature group
feature_group = get_feature_store().get_or_create_feature_group(FEATURE_GROUP_PREDICTIONS_METADATA)


feature_group.insert(predictions, write_options={"wait_for_jobs": False})
